flask_api/requirements.txt
flask
flask-cors
python-dotenv
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1